requirements.txt
Werkzeug==0.15.4
requests2==2.16.0
Flask_RESTful==0.3.7
PyMySQL==0.9.3
Flask==1.0.2
Flask_Cors==3.0.8
PyJWT==1.7.1
Requests==2.22.0
